refactor: routieTime 로직 useRoutieTime 훅으로 분리
시작, 종료 시간에 기본 값을 추가했습니다
