+++ frontend/src/domains/routie/hooks/useRoutieValidate.ts
@@ -1,7 +1,6 @@
 import { useCallback, useMemo, useState } from 'react';
 
 import { useSessionStorage } from '@/@common/hooks/useSessionStorage';
-import { getCombineDateTime } from '@/@common/utils/format';
 
 import { getRoutieValidation } from '../apis/routie';
 import {
@@ -11,6 +10,8 @@ import {
   WaitingReason,
   InvalidRoutiePlace,
 } from '../types/routie.types';
+
+import useRoutieTime from './useRoutieTime';
 
 export interface UseRoutieValidateReturn {
   isValidateActive: boolean;
@@ -43,11 +44,8 @@ const useRoutieValidate = (): UseRoutieValidateReturn => {
     'isValidateActive',
     true,
   );
-  const [routieTime, setRoutieTime] = useSessionStorage('routieTime', {
-    date: '',
-    startTime: '',
-    endTime: '',
-  });
+  const { routieTime, handleTimeChange, combineDateTime, emptyDate } =
+    useRoutieTime();
   const [invalidResult, setInvalidResult] =
     useState<ValidationResultType | null>(null);
   const [validationStatus, setValidationStatus] =
@@ -122,20 +120,6 @@ const useRoutieValidate = (): UseRoutieValidateReturn => {
   const handleValidateToggle = useCallback(() => {
     setIsValidateActive(!isValidateActive);
   }, [isValidateActive, setIsValidateActive]);
-
-  const combineDateTime = useMemo(() => {
-    return getCombineDateTime(routieTime);
-  }, [routieTime]);
-
-  const handleTimeChange = useCallback(
-    (field: string, value: string) => {
-      setRoutieTime({
-        ...routieTime,
-        [field]: value,
-      });
-    },
-    [routieTime, setRoutieTime],
-  );
 
   const validateRoutie = useCallback(
     async (movingStrategy: string, placeCount?: number) => {
